requirements.txt
flask
zeroc-ice
beautifulsoup4
requests
Pillow
